refactor(input): main -> core::input
