+++ src/core/mod.rs
@@ -3,3 +3,4 @@ pub const VERSION: &str = "0.0.1";
 pub mod commands;
 pub mod fs;
 pub mod lua;
+pub mod input;
